♻️ Refactor views into composable files 🔧 Fix SSL redirect issue 🔥 Remove `.html` on endpoints
