+++ landing_page_app/templates/home.html
@@ -1,16 +1,16 @@
 {% extends "base.html" %}
 
 {% block pageTitle %}
-Home
+    Home
 {% endblock %}
 
 {% block content %}
-<h1 class="govuk-heading-xl">Operations-Engineering Landing Page PoC</h1>
+    <h1 class="govuk-heading-xl">Operations-Engineering Landing Page PoC</h1>
 
-<p class="govuk-body">
-    This website is a PoC for the Operations-Engineering Team.
-</p>
+    <p class="govuk-body">
+        This website is a PoC for the Operations-Engineering Team.
+    </p>
 
-<p class="govuk-body"><a href="join-github.html">Join Github</a></p>
+    <p class="govuk-body"><a href="join-github">Join Github</a></p>
 
 {% endblock %}
