+++ landing_page_app/templates/thank-you.html
@@ -1,16 +1,17 @@
 {% extends "base.html" %}
 
 {% block pageTitle %}
-Thank you
+    Thank you
 {% endblock %}
 
 {% block content %}
-<h1 class="govuk-heading-xl">Thank you</h1>
+    <h1 class="govuk-heading-xl">Thank you</h1>
 
-<p class="govuk-body">
-  If you require further assistance contact us on the Operations-Engineering team Slack channel <a href="https://mojdt.slack.com/archives/C01BUKJSZD4" class="govuk-link">#ask-operations-engineering</a>.
-</p>
+    <p class="govuk-body">
+        If you require further assistance contact us on the Operations-Engineering team Slack channel <a
+            href="https://mojdt.slack.com/archives/C01BUKJSZD4" class="govuk-link">#ask-operations-engineering</a>.
+    </p>
 
-<a href="/home" class="govuk-back-link">Back</a>
+    <a href="/" class="govuk-back-link">Back</a>
 
 {% endblock %}
